now signed in after reset and register
