+++ src/api/config/middlewares/authentication.js
@@ -1,35 +1,6 @@
 'use strict';
 
-const passport = require('passport');
-const User = require('../../app/models/user');
-const LocalStrategy = require('passport-local');
-
-const localOptions = { usernameField: 'email' };
-
-const localSignIn = new LocalStrategy(localOptions, (email, password, done) => {
-  User.findOne({ email: email.toLowerCase() })
-    .then((user) => {
-      if (!user) { return done(null, false); }
-
-      user.comparePassword(password)
-        .then((isMatch) => {
-          if (!isMatch) { return done(null, false); }
-
-          return done(null, user);
-        });
-    })
-    .catch(done);
-});
-
-passport.use(localSignIn);
-
-passport.serializeUser(function(user, done) {
-  return done(null, user);
-});
-
-passport.deserializeUser(function(user, done) {
-  return done(null, user);
-});
+const passport = require('../passport');
 
 // This is because idk the best way of testing a user being signed in. Not real code, really.
 // It'll be gone soon.
